rewrite minter function to ts
