--- amplify/backend/function/fnArcaneMint/src/DatabaseInterface.js
@@ -1,20 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.Rarity = exports.Affinity = void 0;
-var Affinity;
-(function (Affinity) {
-    Affinity["Light"] = "Light";
-    Affinity["Dark"] = "Dark";
-    Affinity["Fire"] = "Fire";
-    Affinity["Water"] = "Water";
-    Affinity["Earth"] = "Earth";
-    Affinity["Wind"] = "Wind";
-})(Affinity = exports.Affinity || (exports.Affinity = {}));
-var Rarity;
-(function (Rarity) {
-    Rarity["common"] = "common";
-    Rarity["uncommon"] = "uncommon";
-    Rarity["rare"] = "rare";
-    Rarity["secret"] = "secret";
-    Rarity["legendary"] = "legendary";
-})(Rarity = exports.Rarity || (exports.Rarity = {}));
