+++ amplify/backend/function/fnArcaneMint/src/Environment.ts
@@ -1,6 +1,11 @@
 import * as dotenv from 'dotenv';
 dotenv.config();
 
+/**
+ * Helper function to get environment variables conforming to string type.
+ * @param key environment variable name
+ * @returns environment variable value
+ */
 export function extractStringEnvVar(key: string): string {
     let result = process.env[key];
     if(result === undefined) {
@@ -10,6 +15,11 @@ export function extractStringEnvVar(key: string): string {
     return result;
 }
 
+/**
+ * Helper function to get environment variables conforming to number type.
+ * @param key environment variable name
+ * @returns environment variable value
+ */
 export function extractNumberEnvVar(key: string): number {
     let result = Number(process.env[key]);
