--- amplify/backend/function/fnArcaneMint/src/Typescript/DatabaseInterface.ts
@@ -1,77 +1,0 @@
-export type Familiar = {
-    _id: number,
-    familiarId: number,
-    name: string,
-    description: string,
-    image_url: string,
-    image: string,
-    affinity: Affinity,
-    HP: number,
-    MP: number,
-    attack: number,
-    defense: number,
-    arcane: number,
-    speed: number,
-    ability_1: string,
-    ability_2: string,
-    ability_3: string,
-    ability_4: string,
-    rarity: Rarity,
-    generation: number
-    meta?: {
-        status?: string,
-        mint_timestamp?: string,
-        origin_owner?: string,
-        minted?: number,
-        limit?: number
-    }
-}
-
-export type User = {
-    _id: string,
-    total_minted: number,
-    register_timestamp: string,
-    last_active_timestamp: string,
-    saveData: {
-        name: string,
-        level: number,
-        items: Array<string>,
-        locations: Object,
-        progress: Array<number>
-    }
-    blacklist: boolean,
-    reason: string
-}
-
-export enum Affinity {
-    Light = "Light",
-    Dark = "Dark",
-    Fire = "Fire",
-    Water = "Water",
-    Earth = "Earth",
-    Wind = "Wind"
-}
-
-export enum Rarity {
-    common = "common",
-    uncommon = "uncommon",
-    rare = "rare",
-    secret = "secret",
-    legendary = "legendary",
-}
-
-export interface DBInterface {
-    // Initializer
-    init(): Promise<void | undefined>,
-    isInitialized(): boolean,
-    // Getters
-    getFamiliarByID(tokenId: number): Promise<Familiar | undefined>,
-    getPendingMints(): Promise<Array<Familiar> | undefined>,
-    getMintStatus(tokenId: number): Promise<string | undefined>,
-    getTemplatesByTier(tier: Array<Rarity>): Promise<Array<Familiar> | undefined>,
-    getUserByAddress(address: string): Promise<User | undefined>,
-    // Setters
-    updatePendingMints(tokens: Array<Familiar>): Promise<boolean | undefined>,
-    registerNewFamiliar(template: Familiar, user: User): Promise<boolean | undefined>
-}
-
